+++ src/amqp/vhosts/index.ts
@@ -1,6 +1,7 @@
-import amqplib, { Options, Channel, Connection } from 'amqplib';
+import amqplib, { Options, Channel, Connection, ConfirmChannel } from 'amqplib';
 import { logger } from '../../logger';
 import { BufferConverter } from '../buffer-converter';
+import { AppContainer } from '../../container';
 
 export interface RabbitMQConfig {
   protocol: string;
@@ -8,42 +9,78 @@ export interface RabbitMQConfig {
   port: number;
   username: string;
   password: string;
+  consumerPrefetch: number;
+  producerPrefetch: number;
 }
+
+export type InitOptions = {
+  startConsumers: boolean;
+  startProducers: boolean;
+};
 
 export abstract class RabbitMQ {
   protected connection!: Connection;
-  protected channel!: Channel;
+
+  protected producerChannel!: ConfirmChannel;
+
+  protected consumerChannel!: Channel;
+
   protected readonly config: RabbitMQConfig;
+
   readonly vHost: string;
+
+  protected initOptions?: InitOptions;
+
+  protected defaultInitOptions = {
+    startConsumers: true,
+    startProducers: true,
+  };
 
   constructor(vHost: string, config: RabbitMQConfig) {
     this.config = config;
     this.vHost = vHost;
   }
 
-  async init(): Promise<void> {
+  startConsumers(container: AppContainer): void {}
+
+  async init(
+    container: AppContainer,
+    initOptions: InitOptions = this.defaultInitOptions
+  ): Promise<void> {
+    this.initOptions = initOptions;
+
     try {
       this.connection = await amqplib.connect(this.connectionConfig());
       logger.info(`RabbitMQ connection established on vhost - ${this.vHost}
         `);
-      this.handleOnError();
-      this.channel = await this.connection.createChannel();
+      this.handleOnError(container);
+
+      if (initOptions.startProducers) {
+        this.producerChannel = await this.connection.createConfirmChannel();
+        this.producerChannel.prefetch(this.config.producerPrefetch);
+      }
+
+      if (initOptions.startConsumers) {
+        this.consumerChannel = await this.connection.createChannel();
+        this.consumerChannel.prefetch(this.config.consumerPrefetch);
+        this.startConsumers(container);
+      }
     } catch (err) {
       logger.error(
         `Error connecting RabbitMQ to virtual host ${this.vHost} : ${err}`
       );
-      this.reconnect();
+      this.reconnect(container);
     }
   }
 
-  send(
+  async send(
     exchange: string,
     routingKey: string,
     message: object,
     additionalParams?: Options.Publish
-  ): void {
+  ): Promise<void> {
     try {
-      this.channel.publish(
+      await this.publish(
         exchange || '',
         routingKey,
         BufferConverter.converter(message),
@@ -52,6 +89,25 @@ export abstract class RabbitMQ {
     } catch (err) {
       throw Error(`Error Posting Message to RabbitMQ Server - cause ${err}`);
     }
+  }
+
+  protected async publish(
+    exchange: string,
+    routingKey: string,
+    message: Buffer,
+    options?: Options.Publish
+  ) {
+    return new Promise((resolve, reject) => {
+      this.producerChannel.publish(
+        exchange,
+        routingKey,
+        message,
+        options,
+        (err, ok) => {
+          err ? reject(err) : resolve(ok);
+        }
+      );
+    });
   }
 
   protected connectionConfig(): Options.Connect {
@@ -65,22 +121,20 @@ export abstract class RabbitMQ {
     };
   }
 
-  protected handleOnError() {
+  protected handleOnError(container: AppContainer) {
     this.connection.on('blocked', reason => {
       logger.error(`Connection blocked because of ${reason}`);
     });
-    this.connection.on('close', () => this.reconnect());
-    this.connection.on('error', () => this.reconnect());
+    this.connection.on('close', () => this.reconnect(container));
+    this.connection.on('error', () => this.reconnect(container));
   }
 
-  protected reconnect(): void {
-    delete this.channel;
-    delete this.connection;
+  protected reconnect(container: AppContainer): void {
     logger.warn(
       `Trying to connect to rabbitmq on virtual host ${this.vHost} in 5 seconds`
     );
     setTimeout(() => {
-      this.init();
+      this.init(container, this.initOptions);
     }, 5000);
   }
 }