+++ src/http/controllers/controller.ts
@@ -2,5 +2,6 @@ import { RouteConfig } from '../decorators';
 
 export abstract class BaseController {
   path?: string;
+
   routeConfigs?: RouteConfig[];
 }