+++ src/logger.ts
@@ -1,33 +1,23 @@
-import winston from 'winston';
-import momentTimezone from 'moment-timezone';
+import { init } from '@somosphi/logger';
+import bformat from 'bunyan-format';
+import * as dotenv from 'dotenv';
 
-const {
-  combine, splat, colorize, printf,
-} = winston.format;
+dotenv.config();
 
-export const customFormat = printf((info) => {
-  const { level, message, ...data } = info;
-  const keys = Object.keys(data);
-  const msgsToConcat = [message];
-
-  if (keys.length) {
-    const metadata: any = {};
-    keys.forEach((key) => {
-      metadata[key] = data[key];
-    });
-    msgsToConcat.push(JSON.stringify(metadata));
-  }
-
-  return `[${momentTimezone(info.timestamp)
-    .utc()
-    .format('YYYY-MM-DD HH:mm:ss')}] ${info.level}: ${msgsToConcat.join(' ')}`;
+const formatOut = bformat({
+  outputMode: process.env.LOGGER_BEAUTIFY ? 'short' : 'bunyan',
 });
-
-export const logger = winston.createLogger({
-  format: combine(splat(), colorize(), customFormat),
-  transports: [
-    new winston.transports.File({ filename: 'error.log', level: 'error' }),
-    new winston.transports.File({ filename: 'combined.log' }),
-    new winston.transports.Console({}),
+const { Logger, ExpressLogger, AxiosLogger } = init({
+  PROJECT_NAME: 'node-ts-seed',
+  // @ts-ignore
+  LOG_LEVEL: process.env.LOGGER_LEVEL || 'info',
+  STREAMS: [
+    {
+      stream: formatOut,
+    },
   ],
 });
+
+export const logger = Logger;
+export const expressLogger = ExpressLogger;
+export const axiosLogger = AxiosLogger;