+++ test/amqp/vhosts/index.test.ts
@@ -11,14 +11,18 @@ const connConfig = {
   protocol: 'test',
   port: 3000,
   vhost: 'test',
-}
+};
+
 const rabbitMQConfig: RabbitMQConfig = {
   protocol: '123',
   host: '123',
   port: 123,
   username: 'user',
   password: 'pass',
-}
+  consumerPrefetch: 10,
+  producerPrefetch: 10,
+};
+
 const sandbox = sinon.createSandbox();
 let clock: any;
 
@@ -31,7 +35,6 @@ let fnConnectionOn: SinonStub;
 let fnInit: SinonStub;
 
 describe('RabbitMQ', () => {
-
   beforeEach(() => {
     clock = sandbox.useFakeTimers();
     fnConnectionConfig = sandbox.stub().returns(connConfig);
@@ -42,25 +45,29 @@ describe('RabbitMQ', () => {
     fnInit = sandbox.stub();
 
     connection = {
-      createChannel: sandbox.stub(),
+      createChannel: sandbox.stub().returns({
+        prefetch: sandbox.stub(),
+      }),
+      createConfirmChannel: sandbox.stub().returns({
+        prefetch: sandbox.stub(),
+      }),
     };
-    sandbox.stub(amqplib, 'connect')
+    sandbox
+      .stub(amqplib, 'connect')
       // @ts-ignore
       .callsFake(() => connection);
-
   });
 
   describe('init', () => {
     class TestVHost extends RabbitMQ {
-      constructor(vHost: string, config: RabbitMQConfig) {
-        super(vHost, config);
-      }
       connectionConfig() {
         return fnConnectionConfig();
       }
+
       handleOnError() {
         fnHandleOnError();
       }
+
       reconnect() {
         fnRecconect();
       }
@@ -69,20 +76,62 @@ describe('RabbitMQ', () => {
     it('should connect, handleErros and create channel', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
 
-      await testVhost.init();
+      // @ts-ignore
+      await testVhost.init({});
+
       // @ts-ignore
       sandbox.assert.calledWith(amqplib.connect, connConfig);
       sandbox.assert.calledOnce(fnConnectionConfig);
       sandbox.assert.calledOnce(fnHandleOnError);
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
       sandbox.assert.calledOnce(connection.createChannel);
     });
+
     it('should throw error', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
 
       fnConnectionConfig.throws();
-      await testVhost.init();
+
+      // @ts-ignore
+      await testVhost.init({});
 
       sandbox.assert.calledOnce(fnRecconect);
+    });
+
+    it('shoud start producers and consumers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+
+      // @ts-ignore
+      await testVhost.init({});
+
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
+      sandbox.assert.calledOnce(connection.createChannel);
+    });
+
+    it('shoud start only producers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+      const initOptions = {
+        startConsumers: false,
+        startProducers: true,
+      };
+      // @ts-ignore
+      await testVhost.init({}, initOptions);
+
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
+      sandbox.assert.notCalled(connection.createChannel);
+    });
+
+    it('shoud start only consumers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+      const initOptions = {
+        startConsumers: true,
+        startProducers: false,
+      };
+      // @ts-ignore
+      await testVhost.init({}, initOptions);
+
+      sandbox.assert.notCalled(connection.createConfirmChannel);
+      sandbox.assert.calledOnce(connection.createChannel);
     });
   });
 
@@ -90,49 +139,37 @@ describe('RabbitMQ', () => {
     class TestVHost extends RabbitMQ {
       constructor(vHost: string, config: RabbitMQConfig) {
         super(vHost, config);
-        this.channel = {
+        this.producerChannel = {
           // @ts-ignore
-          publish: (
-            exchange,
-            routingKey,
-            message,
-          ) => {
+          publish: (exchange, routingKey, message) => {
             fnPublish(exchange, routingKey, message);
           },
-        }
+        };
       }
     }
 
     it('should call publish method', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
 
-      testVhost.send(
-        'test',
-        'routingKey',
-        { message: 'test' },
-      );
+      testVhost.send('test', 'routingKey', { message: 'test' });
       sandbox.assert.calledWith(
         fnPublish,
         'test',
         'routingKey',
-        BufferConverter.converter({ message: 'test' }),
+        BufferConverter.converter({ message: 'test' })
       );
     });
 
-    it('should throw error', () => {
+    it('should throw error', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       fnPublish.throws();
 
-      expect(
-        () => {
-          testVhost.send(
-            'test',
-            'routingKey',
-            { message: 'test' },
-          );
-        }
-      ).to.throws();
-
+      try {
+        await testVhost.send('test', 'routingKey', { message: 'test' });
+        sandbox.assert.fail('should throwed error');
+      } catch (error) {
+        expect(error).to.be.instanceOf(Error);
+      }
     });
   });
 
@@ -146,11 +183,14 @@ describe('RabbitMQ', () => {
             fn();
             fnConnectionOn(action, fn);
           },
-        }
+        };
       }
+
       handleOnError() {
-        super.handleOnError();
+        // @ts-ignore
+        super.handleOnError({});
       }
+
       reconnect() {
         fnRecconect();
       }
@@ -170,18 +210,9 @@ describe('RabbitMQ', () => {
     it('should call connection.on with args', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       testVhost.handleOnError();
-      sandbox.assert.calledWith(
-        fnConnectionOn,
-        'blocked',
-      );
-      sandbox.assert.calledWith(
-        fnConnectionOn,
-        'close',
-      );
-      sandbox.assert.calledWith(
-        fnConnectionOn,
-        'error',
-      );
+      sandbox.assert.calledWith(fnConnectionOn, 'blocked');
+      sandbox.assert.calledWith(fnConnectionOn, 'close');
+      sandbox.assert.calledWith(fnConnectionOn, 'error');
     });
   });
 
@@ -194,16 +225,21 @@ describe('RabbitMQ', () => {
         // @ts-ignore
         this.connection = { test: 123 };
       }
+
       // @ts-ignore
       init() {
         fnInit();
       }
+
       reconnect() {
-        super.reconnect();
+        // @ts-ignore
+        super.reconnect({});
       }
-      getChannel() {
-        return this.channel;
+
+      getConsumerChannel() {
+        return this.consumerChannel;
       }
+
       getConnection() {
         return this.connection;
       }
@@ -212,13 +248,9 @@ describe('RabbitMQ', () => {
     it('should delete channel prop', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       testVhost.reconnect();
-      expect(testVhost.getChannel()).to.be.undefined;
+      expect(testVhost.getConsumerChannel()).to.be.undefined;
     });
-    it('should delete connection prop', () => {
-      const testVhost = new TestVHost('test', rabbitMQConfig);
-      testVhost.reconnect();
-      expect(testVhost.getConnection()).to.be.undefined;
-    });
+
     it('should call init method after 5000ms', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       testVhost.reconnect();